+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ## [Unreleased] - ReleaseDate
 
+### Added
+
+- Documentation for the library
+
 ### Fixed
 
 - Allow cleanup when interrupting with Ctrl-C or SIGINT. This will allow to send end messages later